src/requirements/base.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
Faker==23.2.1
flower==2.0.1
frozenlist==1.4.1
humanize==4.9.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
multidict==6.0.5
pillow==10.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
tornado==6.4
twilio==8.13.0
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
pyotp==2.9.0
drf-extensions==0.7.1
pre-commit==3.6.2
python-dotenv==1.0.1
django-debug-toolbar==4.3.0
django-split-settings==1.2.0
rich==13.7.0
pip-lock==2.11.0
ipython==8.22.1
django-rich==1.8.0
flake8-pyproject==1.2.3
flake8==7.0.0
factory-boy==3.3.0
django-linear-migrations==2.12.0
django-version-checks==1.2.0
